test(e2e): extract shared mobile terminal helpers
tapTerminalTab, switchToTerminalPanel, waitForShellReady and
readTerminalBuffer were duplicated verbatim across the
mobile-terminal-keybar and mobile-terminal-scroll specs (~35 lines). Move
them into mobile-terminal-helpers.ts and import from both so future changes
to the connect/readiness path land in one place.

Addresses review feedback from Claude and cubic on the previous commit.
